Amalgamate the sources
